pyproject.toml
[project]
name = "univer-by"
version = "0.1.5"
description = ""
authors = [
    {name = "daniil",email = "bialetskidaniil@gmail.com"},
    {name = "daniel70_andreenko",email = "<d4nilandreenko@yandex.by>"},
    {name = "sered",email = "<seredinskii30@gmail.com>"}
]
license = {text = "MIT license"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-crispy-forms (>=2.3,<3.0)",
    "crispy-bootstrap4 (>=2024.10,<2025.0)",
    "fontawesomefree (>=6.6.0,<7.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "univer"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
